Abandon the references approach, switch to Hashmaps and ids instead
It was never gonna work. As explained in https://stackoverflow.com/a/32300133
> Some crates provide ways of representing this case, but they require
> that the base address never move. This rules out mutating vectors,
> which may cause a reallocation and a move of the heap-allocated
> values.

So for now we will simply keep the keys that uniquely identify a value
in the appropriate HashMap. That does mean passing around the whole
DataTable whenever we need to materialize relational data, but at
least it works and is fast.