check api response for internal reference consistency
Since the API returns somewhat random data, it is necessary to check
if all references, e.g. from a town to the towns player, are correct.
Sometimes the API will return results where the player is no longer
in the player table, but the towns table still has references to that
player.